requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
djangorestframework-jwt
pillow
gunicorn
whitenoise